requirements.txt
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.1
click==7.1.2
dj-database-url==0.5.0
Django==2.2.14
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.1
django-registration-redux==2.8
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
views==0.3
Werkzeug==1.0.1
whitenoise==5.2.0
